made changes as per review comment
